Fix grapple throwing in midair
Fixed a bug where the rope of the grappling hook did not stay attached to the Player object when the player is moving.
Replaced TrailRenderer with a LineRenderer whose elements update at the beginning of each FixedUpdate() loop to represent the player and hook position.
Added private fields holding the player and hook x and y positions and removed some unused code.